thumbnail service in progress

+++ controllers/apiControllers/apiProductsController.js
@@ -1,6 +1,6 @@
 import Product from '../../models/Product.js'
 import createError from 'http-errors'
-import { __dirname } from '../../lib/utils.js'
+import { __dirname, fileExists } from '../../lib/utils.js'
 import fs from 'node:fs/promises'
 import path from 'node:path'
 export async function apiProductsList (req, res, next) {
@@ -110,10 +110,21 @@ export async function apiProductDelete (req, res, next) {
       return next(createError(403, 'Forbidden'))
     }
 
-    if (product.image) {
-      const route = path.join(__dirname, '..', 'public', 'uploads')
+  if (product.image) {
+    const route = path.join(__dirname, '..', 'public', 'uploads')
+    const routeThum = path.join(__dirname, '..', 'public', 'uploads', 'thumbnails')
+    const [existImage, existThum] = await Promise.all([
+      fileExists(route, product.image),
+      fileExists(routeThum, product.image)
+    ])
+
+    if (existImage){
       await fs.unlink(path.join(route, product.image))
     }
+    if (existThum){
+      await fs.unlink(path.join(routeThum, product.image))
+    }
+  }
 
     await Product.deleteOne({ _id: productId })
